Rewrite apiclient tests to avoid dmutils.request_id dependency
Tests for request headers were using the `dmutils.request_id` module
to set up the Request class with `.request_id` attribute. Since the
test is supposed to check that API client is setting the header on
the API request, we can configure the Flask request object directly
instead of relying on `request_id.init_app`.