Always send "updated_by" field instead of "update_details"
Now that API supports both "updated_by" and "update_details" keys
we can switch the apiclient to always send the same payload.

Adds helper methods to add user info to the request data for post,
put and delete requests.

Changes data API client methods to use the new helpers so that update
data is always sent in the same format.