Rename `signer_details` to `agreement_details`
We are expanding the scope of what the *_details JSON field will
store to do with signing framework agreements.  For example,
not all stored values will be about the supplier who signs it
 -- we are also planning to save the name/role of the
countersigner.

With this in mind, we are generalising the name of the column in
the database.  This is easy to do now because no one is using
it yet.