Kill the import_* wrapper methods
Now that we have a better way of doing data migrations onto local
dev machines, we can obliterate the supplier/service import
routes.

The only place these methods are used is in the import script
found in the API. A subsequent pull request on the API will remove
the script as well as the routes that actually do the imports.