HTTPTemporaryError now wraps 502 errors
Previously, only getting back 503 or 504 errors would cause the
`find_{model}_iter` function to retry the failing call.
Now we've added 502 errors as well.

This is to mitigate a problem we're experiencing with exporting
brief responses.  It's happened a few times that the
 `find_brief_responses_iter` function gets through about ~190
pages of brief responses and then fails upon getting back a 502
error.

Manually testing getting back brief responses fails intermittently
but not reliably so we're going to make the API client more
resilient while we're investigating the problem.

This is a short-term-type solution so that we can complete the
data exports, but it doesn't actually do anything to address the
failures in the API.