Add `unlock_user` argument flag to `.update_user_password`
Allows unlocking the user and resetting the failed login count when
the password is changed.

Since password update is the same endpoint as the user update it makes
sense to do both changes in one API request to avoid any inconsistencies
that could be caused by one of the requests failing.

Updater details are the last argument for most API client methods, so
I'm adding a new argument before `updater`, making this a breaking
change. This shouldn't be an issue in practice since the method is
currently only used by the user-frontend.