Alias `get_direct_award_project_services_iter` to `find_`
The ModelTrawler class in the scripts repo uses the `find_<model>_iter`
pattern to match the iter methods in the api client and use them to
fetch data from the api.

We need to use the `get_direct_award_project_services_iter` for data,
and instead of some nasty mangling to the ModelTrawler class, it's
easier to alias the method here.