Index using supplied URL
Previously the `index` method took and index name and an object id and
then contructed the URL from that.

To construct the url it used another method, `_url` which has `services`
hardcoded as the object type. If we changed this to take an object type
instead of just hardcoded `services` it has knock on implications for
frontend app which ALSO use this method, via `get_search_url`.

By defining the url's and appending the object id to the end of them,
and passing that to the `index` method we can get around it.