Give `object_type` to `index`
It was decided that it would be more obvious what was happening if the
search-api client index method took the object type for the document
being indexed, and created the url using it.

The url is being constructed in the method instead of using the `_url`
method, as it's less work than giving `_url` a default, and is more
obvious.