Allow `make_iter_method` accept more than one model name
Rather than creating a temporary custom `search_services_from_url_iter`
method, to allow for the api to accept `documents` as well as `services`
when locking saved search project, this allows `make_iter_method` to
accept more than once model name.

The first model name found in the *args list that is present as a top
level key in the response is used. The rest are discarded.

This is to allow us to update the search-api without breaking other
things. Once the search-api is updated the extra model name can be
removed. It might be useful to keep the functionality on
`make_iter_method` though.