Allow iter methods to not find `links` key in JSON response
As part of a PR to standardise the API responses (
https://github.com/alphagov/digitalmarketplace-api/pull/693), we
may no longer be including `links` dicts as part of an endpoints
response.

An endpoint may change from returning a `paginated_result_response`
to returning a `list_result_response` based on the query params
that have been provided.

We discovered that we have iter methods such as the
`find_services_iter` which will break when passing in a supplier
ID paramater as they now return a standardised `list_result_response`
which no longer includes `links`. However the `iter` methods at
the moment rely on getting a `links` key in their result.

Changing our API client to be permissive feels like the easiest
thing to do which won't have any functionality changes, especially
as it is likely that we were only originally returning an empty
list of links from the API so we wouldn't break the API client
rather than providing extra useful information in our responses.