Move API stubs to dmutils
 ## Summary
The API stubs require some formatters that are declared in utils. We
could pull utils into this repository but they seem to be cleanly
divided at the moment and I'm reticent to start having them depend on
each other.

The stubs are also definitely a utility, so they feel better placed in
dmutils - even though they are related to API responses.

 ## Ticket
https://trello.com/c/BVPgZRZc/65-store-datetimes-relating-to-the-framework-lifecycle-on-the-api