Pin to Flask==0.10.1
 ## Summary
Our frontend apps all depend on Flask v0.10.1 specifically, but this
repository doesn't. Flask released v1.0.0 over the weekend and this is
trying to pull it in now, which breaks things. Let's pin this dependency
to the specific version and review if/how/when/whether we should upgrade
to Flask v1.0.0