Add retry to base api client requests
This was heavily influenced by this blog post: https://www.peterbe.com/plog/best-practice-with-retries-with-requests

This will allow the api client to retry requests which raise connection
errors, as well as requests that return certain 500 error codes.

For the tests, I've had to reach quite a long way in to the requests and
urllib3 modules. This is because response_mock (rmock fixture) mocks the
adapter that requests uses. Doing this means that we never actually
retry anything. I was torn between doing this potentially quite fragile
testing method, or just testing that Retry was called correctly when a
request was made. I am very open to changing.