Add get_buyer_email_domains methods
We've had a request from CCS to see the list of approved email address
domains for buyers. In future we might also want to add an admin
frontend view that displays this info.

This commit adds a method to grab a page from the /buyer-email-domains
api endpoint, and also an iter method so that you can easily fetch all
values in one go.