Add default `updated_by` property to BaseApiClient
This is mainly meant as a quality-of-life improvement for devs using the
API client shell in digitalmarketplace-scripts; instead of having to
remember to add `updater` or `user` to each method call and typing their
email address every time, they can now set it once when constructing the
API client object. It might also help with simplifying writing scripts.

This commit adds a new property to BaseApiClient, `_user`, which can be
`None` or a string. If it is a string this will be used for `updated_by`
when using a method that makes a change using the API. The caller can
still use the user argument for the method call if they want to, and
this will supersede the value of the property, but now they don't have
to. If they haven't provided the argument in the method call or set the
property, a ValueError will be raised.

This commit should be backwards-compatible.